Move main content into a view
